Added int64_t support

+++ Tools/ToolStateObserver/lsqfilter.h
@@ -1,19 +1,21 @@
 #ifndef LSQFILTER_H
 #define LSQFILTER_H
 
+#include <cinttypes>
+
 struct LsqFilter
 {
   LsqFilter(
     const int filter_shift,
-    const int value_active = 0,
-    const int value_hidden = 0);
+    const int64_t value_active = 0,
+    const int64_t value_hidden = 0);
 
-  int Estimate(const int measurement);
+  int64_t Estimate(const int64_t measurement);
 
   private:
-  int m_val1;
-  int m_val2;
-  int m_filter_shift;
+  int64_t m_val1;
+  int64_t m_val2;
+  int64_t m_filter_shift;
 };
 
 #endif // LSQFILTER_H
